Making `resolvableRows` was slowwww if there are a lot of cells. (thanks @bleakley for reporting)

I did some investigation. Runtime increases proportional to the number of cells made (`rows x columns`).

Creating `resolvablePromise` for each cell is relatively expensive. It creates multiple function contexts, and breaks the type hierarchy in a way that the JS runtime probably _hates_.

I made things better by skipping the call to `wrapPromise` and constructing our own resolvable + wrapped promise in one shot. While this is not a total solution to performance issues, this is an improvement of **over 30%** on a common operation.

## Benchmark

```js
const startTime = performance.now()

const numRows = 100_000
const header = ['id', 'name', 'age', 'city']
asyncRows(Promise.resolve([]), numRows, header)

console.log('made async rows in', performance.now() - startTime)
```

| branch | runtime (ms) |
| - | - |
| master | 579ms +/- 81ms | 
| faster-resolvablePromise | 388ms +/- 38ms |