@platypii 2025-05-19
> We might want to deprecate `wrapPromise` and stop exporting it, no? I think we don't need it anymore anywhere.

`wrapPromise` is still used in the stories and also in hyperparam.app toolcalls. So I think its still useful if you want to provide your own promise to hightable.